if no_cache is used in api_query, set url_path
I *believe* there is a bug where if one specified
no_cache to api_query in CachedDocumentManager, then the resulting
CachedDocument will not have an url_path set as I believe it should.
This one-liner patch fixes that.

Apologies if this was not a bug - my testing environment is not yet set.
Really, the major goal of this is to learn how to create branches and
pull requests.